Rutha workspace

+++ src/main/java/CS544/Controller/CommentController.java
@@ -6,8 +6,11 @@ import CS544.Model.User;
 import CS544.Service.CommentService;
 import CS544.Service.PostService;
 import CS544.Service.UserService;
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -32,8 +35,8 @@ public class CommentController {
         return  commentService.getAll();
     }
 
-    @PostMapping(value = "/add/{postId}/{authorId}", consumes = "application/json")
-    public ResponseEntity<?> add(@Valid @RequestBody Comment comment, BindingResult result, @PathVariable Long postId, @PathVariable Long authorId) {
+    @PostMapping(value = "/add/{postId}", consumes = "application/json")
+    public ResponseEntity<?> add(@Valid @RequestBody Comment comment, BindingResult result, @PathVariable Long postId, HttpServletRequest request) {
         if (result.hasErrors()) {
             // Handle validation errors
             List<String> errors = result.getAllErrors().stream()
@@ -42,11 +45,11 @@ public class CommentController {
 
             return ResponseEntity.badRequest().body(errors);
         }
-
-
-        User user = userService.findOne(authorId);
+        Claims claims = (Claims) request.getAttribute("claims");
+        String userName = claims.getSubject();
+        User user = userService.findByUserName(userName);
         Post post=postService.get(postId);
-       comment.setPost(post);
+        comment.setPost(post);
         comment.setAuthor(user);
         commentService.add(comment);
         return ResponseEntity.status(HttpStatus.CREATED).body(comment);
@@ -62,11 +65,11 @@ public class CommentController {
     @PutMapping("/update/{id}")
     public ResponseEntity<?> update(@RequestBody Comment comment,BindingResult result,@PathVariable Long id ) {
         Comment comment1 = commentService.get(id);
+        if(comment1==null){
+            return ResponseEntity.notFound().build();
+        }
         comment1.setDescription(comment.getDescription());
 
-//        if (id != comment.getId()) {
-//            throw new IllegalArgumentException();
-//        }
         if (result.hasErrors()) {
             // Handle validation errors
             List<String> errors = result.getAllErrors().stream()
@@ -80,9 +83,13 @@ public class CommentController {
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public ResponseEntity<Comment> delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        Comment comment1 = commentService.get(id);
+        if(comment1==null){
+            return ResponseEntity.notFound().build();
+        }
         commentService.delete(id);
-        return ResponseEntity.noContent().build();
+        return ResponseEntity.ok("comment deleted successfully");
     }
 }
